+++ Mods/src/Mod.cpp
@@ -1,7 +1,7 @@
 #include "../Headers/Mod.h"
 
-Mod::Mod(std::string newType, int newRequiredLevel, int newLowerMin, int newLowerMax,
-        int newUpperMin, int newUpperMax, std::string newName){
+Mod::Mod(std::string newType, int newRequiredLevel, float newLowerMin, float newLowerMax,
+        float newUpperMin, float newUpperMax, std::string newName){
 
             type = newType;
             requiredLevel = newRequiredLevel;
